+++ R/utils.R
@@ -1,5 +1,5 @@
 
-#' Adds georeferenced data to downloaded data
+#' Adds georeferenced data to the downloaded data
 #'
 #' @export
 #' @aliases add_geo
@@ -9,6 +9,11 @@
 #'
 add_geo <- function(data, ...){
   level <- attributes(data)$level
+
+  if(level=="cities"){
+    data <- data %>%
+      filter(!is.na(pop))
+  }
   if(level == "brazil"){
     return(data)
   }else{
@@ -25,4 +30,26 @@ add_geo <- function(data, ...){
     return(newdata)
   }
 }
+
+
+
+#' Adds epidemiolocical information (incidence, lethality and mortality rates) to the downloaded data
+#'
+#' @export
+#' @aliases add_epi_info
+#' @param data a data set downloaded using the covid19br::downloadCovid19() function.
+#' @param ... further arguments passed to other methods.
+#' @return the data set with the added epidemiolocical information.
+#'
+
+add_epi_info <- function(data, ...){
+  newdata <- data %>%
+    mutate(
+      incidence = 100000*accumCases/pop,
+      lethality = round(100*(accumDeaths/accumCases), 2),
+      mortality = 100000*accumDeaths/pop
+    )
+  return(newdata)
+}
+