+++ tests/testthat/test_covid19br.R
@@ -9,3 +9,24 @@ cities <- downloadCovid19("cities")
 
 # Downloading world COVID-19 data:
 world <- downloadCovid19("world")
+
+
+
+# adding the geometry to the data:
+regions_geo <- regions %>%
+  filter(date == max(date)) %>%
+  add_geo()
+
+states_geo <- states %>%
+  filter(date == max(date)) %>%
+  add_geo()
+
+cities_geo <- cities %>%
+  filter(date == max(date)) %>%
+  add_geo()
+
+world_geo <- world %>%
+  filter(date == max(date)) %>%
+  add_geo()
+
+