+++ NEWS.md
@@ -4,8 +4,10 @@
 
 # covid19br 0.1.2
 
-- Addition of georeferenced data for Brazil (at city, state and region levels) and world (at country level).
+- Implementation of add_epi_info() function to add epidemiological rates (incidence, mortality and lethatity) to the downloaded COVID19 data.
 
 - Implementation of add_geo() function to add georeferenced data to the downloaded COVID19 data.
+
+- Addition of georeferenced data for Brazil (at city, state and region levels) and world (at country level).
 
 - Lubridate package removed from the Imports field in the DESCRIPTION file.